+++ CHANGES.rst
@@ -10,6 +10,7 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.4.3 (unreleased)
 ------------------
 
+- Mark requests with ``Accept: application/json`` header as ``zope.publisher.interfaces.http.IJSONRequest`` requests.
 - Update dependencies to newest releases.
 - Improve solidity of the ``debugError`` method.
   (`#829 <https://github.com/zopefoundation/Zope/issues/829>`_)